Extract and make DebugInfo collapsable js_only
