+++ assets/src/application/ui/display/index.ts
@@ -4,3 +4,4 @@ export { default as EmptyState } from './EmptyState';
 export * from './EntityIDs';
 export { default as ErrorIndicator } from './ErrorIndicator';
 export { default as LoadingIndicator } from './LoadingIndicator';
+export { default as DebugInfo } from './DebugInfo';
